Allow for empty unprocessed items map.
Annoyingly, `hasUnprocessedItems` only checks if `unprocessedItems` is
null but it can also come back as an empty map.
So `hasUnprocessedItems` will be true but `unprocessedItems.get(tableName)` will be
null which causes a NPE.
This check makes sure the map is not empty. I've added an empty map into
the writeItems test to make sure it works.