DR2-1784 Batch delete items
The call to batchWriteItem can also be used for deleting items. The only
difference is the write request. I've moved the existing writeItem logic
to a private method with a parameter for the write request and the
writeItems and deleteItems methods call this.

I've put another table in the tests so the existing writeItems tests are
now being called with deleteItems as well.